pyproject.toml
[tool.poetry]
name = "vocabgenerator"
version = "0.1.0"
description = ""
authors = ["cambrian-dk <daryl.kohyikai@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydub = "^0.25.1"
selenium = "^4.4.3"
PyYAML = "^6.0"
webdriver-manager = "^3.8.3"
google-cloud-texttospeech = "^2.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
